requirements.txt
# LangChain 
langchain==0.2.11
langchain-core>=0.2.23
langchain-community==0.2.5
langchain-huggingface==0.0.3

# Vector DB 
chromadb==0.5.0

# HF Endpoints
huggingface-hub>=0.24.0
requests>=2.32.0

# Document Loaders
pypdf>=4.0.0
python-docx>=1.1.0
docx2txt>=0.8

# HTML / Web Scraping
beautifulsoup4>=4.12.0
lxml>=4.9.3

# Utils
python-dotenv>=1.0.0
tqdm>=4.66.2
